requirement.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
notebook
mlflow
joblib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
types-PyYAML
lightgbm
xgboost
dagshub
flask
flask-cors
-e .